# Optimize execution loop 