--- a/.github/workflows/test-external.yml
+++ b/.github/workflows/test-external.yml
@@ -9,7 +9,7 @@ on:
 
 jobs:
   test:
-    runs-on: macos-latest
+    runs-on: ubuntu-latest
 
     strategy:
       fail-fast: false
@@ -55,6 +55,12 @@ jobs:
       - name: Install Foundry
         uses: foundry-rs/foundry-toolchain@v1
 
+      - name: Install yices
+        run: |
+          sudo add-apt-repository ppa:sri-csl/formal-methods
+          sudo apt-get update
+          sudo apt-get install yices2
+
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
@@ -67,5 +73,5 @@ jobs:
         run: python -m pip install -e ./halmos
 
       - name: Test external repo
-        run: ${{ matrix.cmd }} -v -st --solver-timeout-assertion 0 --solver-threads 2
+        run: ${{ matrix.cmd }} -v -st --solver-timeout-assertion 0 --solver-threads 6 --solver-command=yices-smt2
         working-directory: ${{ matrix.dir }}