--- a/.github/workflows/test-external.yml
+++ b/.github/workflows/test-external.yml
@@ -73,5 +73,5 @@ jobs:
         run: python -m pip install -e ./halmos
 
       - name: Test external repo
-        run: ${{ matrix.cmd }} -v -st --solver-timeout-assertion 0 --solver-threads 6 --solver-command=yices-smt2
+        run: ${{ matrix.cmd }} -v -st --solver-timeout-assertion 0 --solver-threads 4 --solver-command=yices-smt2
         working-directory: ${{ matrix.dir }}