Completely driven by flamegraphs, I kept squeezing until the fetch-decode-execute loop almost disappeared from flamegraphs (now dominated by other things like calls to `resolve_address_alias`):

Tested on morpho blue, 24.8% faster:

    before: check_fee(bytes4,address) (paths: 38, time: 21.82s (paths: 21.82s, models: 0.00s), bounds: [])
    after:  check_fee(bytes4,address) (paths: 39, time: 16.41s (paths: 16.41s, models: 0.00s), bounds: [])

Avoids duplicating work:

- load pgm[pc] to know the current instruction (doing a chunk lookup + indexing in the code bytevec)
- load pgm[pc+1:pc+1+operand_size] for PUSHn instructions (every time it is executed, which can be expensive)
- load pgm[pc] again to know the next_pc

Instead we do this eagerly once and then remember the result. 


Plus some extra goodies:
- support for MCOPY 
- explicit Message.call_scheme, which caused us to render deploy output in traces at high verbosity when the intent was to just display the number of bytes