@karmacoma-eth 2024-05-22
Tested on Farcaster, 11% faster:

```
# before
[PASS] check_Invariants_PostMigration(bytes4,address) (paths: 641, time: 195.43s (paths: 195.43s, models: 0.00s), bounds: [])

# after
[PASS] check_Invariants_PostMigration(bytes4,address) (paths: 637, time: 173.95s (paths: 173.95s, models: 0.00s), bounds: [])
```